Move api/* to api/ancient/.
Client code is to be expected to include public header files in the form <ancient/Decompressor.hpp>.
Having the header files lie in the same folder structure also in the repository simplifies usage for client code that maintains a copy of the ancient source tree inside its own source tree.
