+++ api/ancient/Decompressor.hpp
